+++ src/main/java/com/qa/ims/utils/Utils.java
@@ -69,7 +69,7 @@ public class Utils {
 			} catch (NumberFormatException nfe) {
 				LOGGER.info("Invalid input - Please enter again");
 			}
-		} while (Day == 0 || Month == 0 || Year == 0 || (Day>31&&Day<1) || (Month>12&&Month<1) || (Year<21));
+		} while (Day == 0 || Month == 0 || Year == 0 || (Day>31 || Day<1) || (Month>12 || Month<1) || (Year<21));
 		return String.valueOf(Day)+"/"+String.valueOf(Month)+"/"+String.valueOf(Year);
 	}
 }
